✨ feat(apiClient): Implement browser and server API client instances
- apiClient.ts를 factory 패턴으로 클라이언트와 서버 로직을 분리
- 클라이언트 컴포넌트에서는 기존과 동일하게 apiClient로 wrapper 사용 가능
- 서버 컴포넌트에서는 apiClientServer을 통해 wrapper 사용 가능
